feat(Alert.js): add Alert.js component
This commit adds the Alert.js component.

This component will display an alert with an alertType and
an alertText.

For example, if the user attempts to make a search without entering
any search criteria, an alert can be shown to the user with an
alertType of 'danger' and an alertText of
'Please enter something to search for'.