feat(Search.js): add Search.js component
This commit adds the Search.js component.

This component will display an input to take in the users search
string and display a 'Search' button the user will press to execute
the search.

TODO: Add props as needed.

TODO: Add call to a function that shows an alert if needed.

TODO: Add call to a function that gets the results from the
user’s search.