feat(App.js): add instance of Search component
This commit adds:

An instance of Search component.

A 'main' element to act as a container for all of the
non-navbar elements (ex. Search) to ensure that they are shown with
the correct width for a given screen size.

This commit deletes:

An unneeded comment pointing out that component level state is used
in App.js.