feat(App.js): use Alert component in App.js
This commit:

Uses the Alert component in App.js.

Adds alert state to App.js.

Adds a comment explaining that getWeatherForcastData() is called
one time as soon as the component loads. This weather data will be
used to provide the user with default weather data as soon as
the app is loaded.