feat(Search.js): Add calls to showAlert() and getWeatherForcastData()
This commit:

Adds a call to showAlert() if the user attempts to search without
inputting any search text.

Adds a call to getWeatherForcastData() if the user attempts to search
after entering search text.

Destructures showAlert and getWeatherForcastData from props so that
showAlert and getWeatherForcastData can be used directly, instead of
having to do props.showAlert or props.getWeatherForcastData.