feat(WeatherHeader.js): add WeatherHeader.js
This commit adds the WeatherHeader component. This component is
used to display the location of the weather data.

For example, if the user searches for 'New York, US', then the
WeatherHeader component will display 'New York, US', so that the user
will understand the weather data is for the location 'New York, US'.