feat(Apps.js): update call to showAlert
This commit updates the call to showAlert that alerts the user if the
location they searched for was not found by the weather app.

The alert now includes the location the user searched for. This way,
if the location was not found due to a typo, it should be more
obvious to the user.