feat(App.css) Add a media query and a css class
This commit:

Adds a media query that executes the contained css when the viewport
is 600px wide or wider.

Adds the singleDayWeatherInfoContainer__fiveDayForecast class. When
the five-day forecast is displayed, this class will display the
contained items in a row if the viewport less than 600px wide and
displays the contained items in a column if the viewport is 600px
wide or greater.

Deletes unneeded property from
singleDayWeatherInfoContainer__currentDayForecast class.

Adds comments distinguishing the mobile css and desktop css for the
ForcastItem class. This should make the css more readable.