feat(App.js): add userClickedWeatherForecastObject
This commit adds the userClickedWeatherForecastObject state with an
initial value of null.

The userClickedWeatherForecastObject state will be used to hold the
weather forecast object that is retrieved when the user clicks the
'details' button for a specific day of the week. This day of the week
is either the current-day forecast or one of the five days of the
five-day forecast.

This weather forecast object will then be used to provide all of the
weather values in the MoreDetailsSingleDayForecast.js component.