feat(App.js): use getUserClickedWeatherForecastObject()
This commit adds the function getUserClickedWeatherForecastObject as a
prop in the FiveDayForecast component and the CurrentDayForecast
component.

Both the FiveDayForecast and CurrentDayForecast components will use the
getUserClickedWeatherForecastObject prop to get the weather forecast
object that the user clicks on and then pass that object back up to
App.js where the userClickedWeatherForecastObject state will be
set to be equal to the weather forecast object that the user clicked on.