feat(ForecastItem.js): update button to Link
This commit:

Updates the 'Details' button to be a Link. This was done because
Links are more properly used to navigate the user to a different
portion of a website (in this case, the more details single-day forcast
page), whereas buttons are more properly used to execute
an action.

NOTE: clicking the 'Details' Link, calls the function
onClickGetUserClickedWeatherForecastObject. This function is used to
get the weather object that the user clicked on in order to provide
weather values for the MoreDetailsSingleDayForecast.js component.

Fixes a typo in instances of the word 'forcast'. The word is now
spelled correctly with an 'e' as in 'forecast'.