fix(FiveDayForecast.js): fix bug with a className
This commit fixes a bug with the conditional omission of the className
'displayFiveDayForecast'.

The code was conditionally omitting the className using the logical
AND (&&) operator. Unfortunately, this was causing an error.

The code now uses the conditional (ternary) operator and sets the
className to undefined if showFiveDayForecastFlag is false. This
prevents the error.