feat: install redux, thunk, and devtools-extension
This commit:

1.) Installs Redux which will be used to manage global state.

Currently, the weather app stores all of the 'app' level state in the
App.js component. This sometimes requires the use of prop drilling to
share state from one component to another.

  For example, the value of the state showFiveDayForecastFlag is
  retrieved from the Navbar.js component, then the state is set in the
  App.js component, then the state is passed down to the
  FiveDayForecast.js component, and then the state is finally passed
  down to the ForecastItem.js component where it is used.

This is a cumbersome way of sharing state between components, which
only becomes more complicated as the number of components is increased.

In contrast, with Redux the global state of the app is stored in
an object tree inside a single store. This enables any global state to
be passed directly to any component without the use of prop drilling.
This will make the weather app easier to maintain and enhance.

2.) Installs Redux-Thunk which is a kind of async middleware that
enables the use of async logic in functions (in this case, within the
action creators of Redux).

3.) Installs Redux-Devtools-Extension which is required to use the
Redux DevTools. The Redux DevTools are used for debuggin Redux.