feat(src): add folder 'actions' & file 'types.js'
This commit adds the folder 'actions' and the file 'types.js'.

The file 'types.js' will be used to contain all of the types
(i.e. actions) that the weather app will do.

NOTE: Currently, the only types listed in this file are SET_ALERT and
REMOVE_ALERT. The other types the weather app will use will be added as
they are needed.