feat(forecastActions.js): add getUserClickedWeatherForecastObjectAction
This commit adds the getUserClickedWeatherForecastObjectAction. This
action will be used to get the specific weather forecast object that
the user clicked on when the user clicks the 'Details' button for a
particular weather forecast day.

Note: The userClickedWeatherForecastObject will be used to set the
value of the state userClickedWeatherForecastObject in the
forecastReducer. Then the value of the state
userClickedWeatherForecastObject will be used to populate the weather
values of the more-details-single-day-forecast page.