fix(WeatherHeader.js): fix PropTypes
This commit fixes the PropTypes for the WeatherHeader.js component by
deleting the PropTypes for 'loading' and 'weatherForecastLocation'
and replacing them with PropTypes for 'forecastReducerStateAsAProp'.

This was done because 'forecastReducerStateAsAProp' is being brought in
from global state as a prop and the values for 'loading' and
'weatherForecastLocation' are destructured from
'forecastReducerStateAsAProp'.