refactor(CurrentDayForecast.js): refactor PropTypes
This commit refactors the PropTypes for the CurrentDayForecast.js
component by deleting the PropTypes for 'fiveDayWeatherForecastArray'
and 'loading', and then replacing them with PropTypes for
'forecastReducerStateAsAProp'.

This was done because 'forecastReducerStateAsAProp' is being brought in
from global state as a prop and the values for
'fiveDayWeatherForecastArray' and 'loading' are destructured from
'forecastReducerStateAsAProp'.