fix(MoreDetailsSingleDayForecast.js): change Link to button
This commit changes the 'Back' button for the
MoreDetailsSingleDayForecast.js component from a react-router Link
component to be an html button element.

This was done because the code was using an onClick event attribute to
programmatically call 'history.goBack();' while keeping the Link
component's 'to' attribute empty. This caused the following warning to
be thrown:

  "Warning: Failed prop type: The prop `to` is marked as required in
  `Link`, but its value is `undefined`."

To correct this, the Link component was changed to a button element
which can use the onClick attribute while not using the unneeded 'to'
attribute.