refactor(forecastActions.js): update map to forEach
This commit refactors forecastActions.js by updating
getWeatherForecastAction to use .forEach() instead of .map() when
iterating over the list of weather objects.

This was done because the code only needs to iterate over the array in
order to find the desired weather data. The code does not need to
return a new / updated array. Therefore the use of .forEach() is the
desired array method.