fix(forecastReducer.js): update GET_WEATHER_FORECAST_LOCATION
This commit no longer sets loading to false in case
GET_WEATHER_FORECAST_LOCATION.

This was necessary because the code is not done loading resources after
finishing with case GET_WEATHER_FORECAST_LOCATION.

Specifically, getWeatherForecastAction dispatches to the reducer twice.
Once for GET_WEATHER_FORECAST_LOCATION and then again for
GET_WEATHER_FORECAST.

Therefore, after the code finishes dispatching to the reducer for
GET_WEATHER_FORECAST_LOCATION, the code needs to continue loading until
it finishes dispatching to the reducer for GET_WEATHER_FORECAST as well.