Fix(App.css): update CSS
This commit updates the text size for the following CSS classes:

.weatherHeader,
.dayOfTheWeek,
.weatherDetails__weatherDescription,
.weatherDetails__weatherTemp,
.weatherDetails__detailsBtn,
.weatherBack__detailsBtn,
.moreWeatherDetails

This was done because FiveDayForcast component was not displaying properly in desktop mode but was clipping off the bottom of the page. I presume this was due to a browser update because this issue was not present when the weather app was originally created and deployed.